Export object_store integration tests
