+++ object_store/src/aws/mod.rs
@@ -409,14 +409,16 @@ impl MultipartStore for AmazonS3 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{client::get::GetClient, tests::*};
+    use crate::client::get::GetClient;
+    use crate::integration::*;
+    use crate::tests::*;
     use hyper::HeaderMap;
 
     const NON_EXISTENT_NAME: &str = "nonexistentname";
 
     #[tokio::test]
     async fn s3_test() {
-        crate::test_util::maybe_skip_integration!();
+        maybe_skip_integration!();
         let config = AmazonS3Builder::from_env();
 
         let integration = config.build().unwrap();
@@ -475,7 +477,7 @@ mod tests {
 
     #[tokio::test]
     async fn s3_test_get_nonexistent_location() {
-        crate::test_util::maybe_skip_integration!();
+        maybe_skip_integration!();
         let integration = AmazonS3Builder::from_env().build().unwrap();
 
         let location = Path::from_iter([NON_EXISTENT_NAME]);
@@ -488,7 +490,7 @@ mod tests {
 
     #[tokio::test]
     async fn s3_test_get_nonexistent_bucket() {
-        crate::test_util::maybe_skip_integration!();
+        maybe_skip_integration!();
         let config = AmazonS3Builder::from_env().with_bucket_name(NON_EXISTENT_NAME);
         let integration = config.build().unwrap();
 
@@ -500,7 +502,7 @@ mod tests {
 
     #[tokio::test]
     async fn s3_test_put_nonexistent_bucket() {
-        crate::test_util::maybe_skip_integration!();
+        maybe_skip_integration!();
         let config = AmazonS3Builder::from_env().with_bucket_name(NON_EXISTENT_NAME);
         let integration = config.build().unwrap();
 
@@ -513,7 +515,7 @@ mod tests {
 
     #[tokio::test]
     async fn s3_test_delete_nonexistent_location() {
-        crate::test_util::maybe_skip_integration!();
+        maybe_skip_integration!();
         let integration = AmazonS3Builder::from_env().build().unwrap();
 
         let location = Path::from_iter([NON_EXISTENT_NAME]);
@@ -523,7 +525,7 @@ mod tests {
 
     #[tokio::test]
     async fn s3_test_delete_nonexistent_bucket() {
-        crate::test_util::maybe_skip_integration!();
+        maybe_skip_integration!();
         let config = AmazonS3Builder::from_env().with_bucket_name(NON_EXISTENT_NAME);
         let integration = config.build().unwrap();
 
@@ -560,7 +562,7 @@ mod tests {
     }
 
     async fn s3_encryption(store: &AmazonS3) {
-        crate::test_util::maybe_skip_integration!();
+        maybe_skip_integration!();
 
         let data = PutPayload::from(vec![3u8; 1024]);
