+++ object_store/src/limit.rs
@@ -273,9 +273,9 @@ impl MultipartUpload for LimitUpload {
 
 #[cfg(test)]
 mod tests {
+    use crate::integration::*;
     use crate::limit::LimitStore;
     use crate::memory::InMemory;
-    use crate::tests::*;
     use crate::ObjectStore;
     use futures::stream::StreamExt;
     use std::pin::Pin;
